Chore: "page.tsx" 복원, inputs.tsx에 사용 예시 작성"
